+++ package.json
@@ -110,6 +110,7 @@
     "verdaccio-memory": "workspace:*"
   },
   "scripts": {
+    "prepare": "husky install",
     "husky:pre-commit": "lint-staged",
     "clean": "pnpm recursive run clean",
     "build": "pnpm recursive run build --filter=!@verdaccio/website",
