+++ test/cli/e2e-npm6/CHANGELOG.md
