+++ test/cli/e2e-npm6/README.md
