+++ test/cli/e2e-npm6/config/_bootstrap_verdaccio.yaml
