+++ test/cli/e2e-npm6/config/default.yaml
