+++ test/cli/e2e-npm6/package.json
@@ -1,0 +1,12 @@
+{
+  "private": true,
+  "name": "@verdaccio/e2e-cli-npm6",
+  "version": "2.0.0-6-next.1",
+  "dependencies": {
+    "@verdaccio/e2e-cli-commons": "workspace:1.0.0",
+    "npm": "latest-6"
+  },
+  "scripts": {
+    "test": "jest"
+  }
+}
