+++ test/cli/e2e-npm6/tsconfig.json
@@ -3,5 +3,10 @@
   "compilerOptions": {
     "rootDir": "./src",
     "outDir": "./build"
-  }
+  },
+  "references": [
+    {
+      "path": "../cli-commons"
+    }
+  ]
 }
