--- test/e2e-cli/.babelrc
@@ -1,3 +1,0 @@
-{
-  "extends": "../../.babelrc"
-}
