--- test/e2e-cli/.eslintrc
@@ -1,7 +1,0 @@
-{
-  "rules": {
-    "no-console": 0,
-    "@typescript-eslint/no-var-requires": 0,
-    "@typescript-eslint/explicit-member-accessibility": 0
-  }
-}
