--- test/e2e-cli/env_babel.js
@@ -1,4 +1,0 @@
-require('@babel/register')({
-  extensions: ['.ts', '.js'],
-});
-module.exports = require('./setup/test_environment');
