--- test/e2e-cli/env_teardown.js
@@ -1,4 +1,0 @@
-require('@babel/register')({
-  extensions: ['.ts', '.js'],
-});
-module.exports = require('./setup/teardown');
