--- test/e2e-cli/package.json
@@ -1,21 +1,0 @@
-{
-  "private": true,
-  "name": "@verdaccio/e2e-cli",
-  "version": "2.0.0-6-next.1",
-  "dependencies": {
-    "debug": "4.3.3",
-    "npm": "7.24.2",
-    "pnpm": "6.27.1",
-    "request": "2.88.0",
-    "yarn": "1.22.17"
-  },
-  "devDependencies": {
-    "get-port": "5.1.1",
-    "wait-on": "^6.0.0",
-    "fs-extra": "10.0.0",
-    "semver": "7.3.5"
-  },
-  "scripts": {
-    "test": "jest --config jest.config.e2e.cli.js"
-  }
-}
