--- test/e2e-cli/projects/basic/index.js
@@ -1,6 +1,0 @@
-module.exports = function () {
-  const message = 'this is a basic project';
-  console.log(message);
-
-  return message;
-};
