--- test/e2e-cli/projects/basic/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "basic-verdaccio",
-  "version": "1.0.0",
-  "description": "this is a basic project",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [
-    "verdaccio",
-    "sample"
-  ],
-  "dependencies": {
-    "verdaccio": "latest"
-  },
-  "author": "Juan Picado <juanpicado19@gmail.com>",
-  "license": "MIT"
-}
