--- test/e2e-cli/projects/scoped_basic/README.md
@@ -1,3 +1,0 @@
-# Simple project
-
-This is a normal readme
