--- test/e2e-cli/projects/scoped_basic/index.js
@@ -1,6 +1,0 @@
-module.exports = function () {
-  const message = 'this is a scoped basic project';
-  console.log(message);
-
-  return message;
-};
