--- test/e2e-cli/test_bk/__partials/npm_commands.ts
@@ -1,13 +1,0 @@
-import { silentNpm } from '../../utils/process';
-
-export function installVerdaccio(verdaccioInstall) {
-  return silentNpm(
-    'install',
-    '--prefix',
-    verdaccioInstall,
-    'verdaccio',
-    '--registry',
-    'http://localhost:6001',
-    '--no-package-lock'
-  );
-}
