--- test/e2e-cli/utils/registry.spec.ts
@@ -1,7 +1,0 @@
-import { addNpmPrefix } from './registry';
-
-describe('registry utils', function () {
-  test('prefix', () => {
-    expect([...addNpmPrefix('foo')]).toEqual('');
-  });
-});
